Unping runtime dependencies
Since `barely_json` is intended to be used as a library, pinning the
runtime dependencies to exact versions is not ideal. Instead, we just
specify the lowest supported version.

We keep the pinning for the dev and test dependencies to ensure that all
versions are the same on local development environments and on the CI.